refactor paths from hyphens to underscores to be more in line with deno
