+++ exec/deno.json
@@ -7,10 +7,10 @@
         "./node": "./node/mod.ts",
         "./command": "./base.ts",
         "./command-base": "./command.ts",
-        "./command-args": "./command-args.ts",
-        "./path-finder": "./path-finder.ts",
+        "./command-args": "./command_args.ts",
+        "./path-finder": "./path_finder.ts",
         "./splat": "./splat.ts",
-        "./split-arguments": "./split-arguments.ts",
+        "./split-arguments": "./split_arguments.ts",
         "./which": "./which.ts",
         "./types": "./types.d.ts",
         "./errors": "./errors.ts"
@@ -20,14 +20,14 @@
         "include": [
             "mod.ts",
             "command.ts",
-            "command-args.ts",
+            "command_args.ts",
             "deno/mod.ts",
             "node/mod.ts",
             "errors.ts",
             "base.ts",
-            "path-finder.ts",
+            "path_finder.ts",
             "splat.ts",
-            "split-arguments.ts",
+            "split_arguments.ts",
             "which.ts",
             "types.d.ts",
             "README.md",
