+++ exec/mod.ts
@@ -61,8 +61,8 @@
  */
 
 export * from "./types.d.ts";
-export * from "./command-args.ts";
+export * from "./command_args.ts";
 export * from "./splat.ts";
-export * from "./split-arguments.ts";
+export * from "./split_arguments.ts";
 export * from "./base.ts";
 export * from "./which.ts";
