+++ exec/path_finder.ts
