+++ fs/deno.json
@@ -8,13 +8,13 @@
         "./constants": "./constants.ts",
         "./base": "./base.ts",
         "./copy": "./copy.ts",
-        "./empty-dir": "./empty-dir.ts",
-        "./ensure-dir": "./ensure-dir.ts",
-        "./ensure-file": "./ensure-file.ts",
-        "./ensure-link": "./ensure-link.ts",
-        "./ensure-symlink": "./ensure-symlink.ts",
+        "./empty-dir": "./empty_dir.ts",
+        "./ensure-dir": "./ensure_dir.ts",
+        "./ensure-file": "./ensure_file.ts",
+        "./ensure-link": "./ensure_link.ts",
+        "./ensure-symlink": "./ensure_symlink.ts",
         "./exists": "./exists.ts",
-        "./expand-glob": "./expand-glob.ts",
+        "./expand-glob": "./expand_glob.ts",
         "./move": "./move.ts",
         "./walk": "./walk.ts"
     },
@@ -32,13 +32,13 @@
             "errors.ts",
             "constants.ts",
             "copy.ts",
-            "empty-dir.ts",
-            "ensure-dir.ts",
-            "ensure-file.ts",
-            "ensure-link.ts",
-            "ensure-symlink.ts",
+            "empty_dir.ts",
+            "ensure_dir.ts",
+            "ensure_file.ts",
+            "ensure_link.ts",
+            "ensure_symlink.ts",
             "exists.ts",
-            "expand-glob.ts",
+            "expand_glob.ts",
             "move.ts",
             "walk.ts",
             "utils.ts",
