+++ fs/empty_dir.ts
