+++ fs/ensure_dir_deno_test.ts
@@ -1,8 +1,8 @@
 // Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
 import { assertEquals, assertRejects, assertThrows } from "@std/assert";
 import * as path from "@std/path";
-import { ensureDir, ensureDirSync } from "./ensure-dir.ts";
-import { ensureFile, ensureFileSync } from "./ensure-file.ts";
+import { ensureDir, ensureDirSync } from "./ensure_dir.ts";
+import { ensureFile, ensureFileSync } from "./ensure_file.ts";
 import { lstat, lstatSync, makeDir, makeDirSync, remove, removeSync, stat, statSync } from "./base.ts";
 
 const moduleDir = path.dirname(path.fromFileUrl(import.meta.url));
