+++ fs/ensure_file_node_test.ts
@@ -2,7 +2,7 @@
 import {} from "./node/shim.ts";
 import { assertRejects, assertThrows } from "@std/assert";
 import * as path from "@std/path";
-import { ensureFile, ensureFileSync } from "./ensure-file.ts";
+import { ensureFile, ensureFileSync } from "./ensure_file.ts";
 import { makeDir, makeDirSync, remove, removeSync, stat, statSync, writeFile, writeFileSync } from "./base.ts";
 
 const moduleDir = path.dirname(path.fromFileUrl(import.meta.url));
