+++ fs/ensure_link_node_test.ts
@@ -3,7 +3,7 @@
 import {} from "./node/shim.ts";
 import { assertEquals, assertRejects, assertThrows } from "@std/assert";
 import * as path from "@std/path";
-import { ensureLink, ensureLinkSync } from "./ensure-link.ts";
+import { ensureLink, ensureLinkSync } from "./ensure_link.ts";
 import {
     lstat,
     lstatSync,
