+++ fs/ensure_symlink_deno_test.ts
@@ -2,7 +2,7 @@
 // TODO(axetroy): Add test for Windows once symlink is implemented for Windows.
 import { assertEquals, assertRejects, assertThrows } from "@std/assert";
 import * as path from "@std/path";
-import { ensureSymlink, ensureSymlinkSync } from "./ensure-symlink.ts";
+import { ensureSymlink, ensureSymlinkSync } from "./ensure_symlink.ts";
 import {
     lstat,
     lstatSync,
