+++ fs/expand_glob_node_test.ts
@@ -2,7 +2,7 @@
 import {} from "./node/shim.ts";
 import { assert, assertEquals, assertStringIncludes } from "@std/assert";
 import { fromFileUrl, join, joinGlobs, normalize, relative } from "@std/path";
-import { expandGlob, type ExpandGlobOptions, expandGlobSync } from "./expand-glob.ts";
+import { expandGlob, type ExpandGlobOptions, expandGlobSync } from "./expand_glob.ts";
 import { cwd } from "./base.ts";
 
 async function expandGlobArray(
