+++ fs/mod.ts
@@ -38,12 +38,12 @@
 export * from "./base.ts";
 export * from "./constants.ts";
 export * from "./copy.ts";
-export * from "./empty-dir.ts";
-export * from "./ensure-dir.ts";
-export * from "./ensure-file.ts";
-export * from "./ensure-link.ts";
-export * from "./ensure-symlink.ts";
+export * from "./empty_dir.ts";
+export * from "./ensure_dir.ts";
+export * from "./ensure_file.ts";
+export * from "./ensure_link.ts";
+export * from "./ensure_symlink.ts";
 export * from "./errors.ts";
 export * from "./move.ts";
 export * from "./walk.ts";
-export * from "./expand-glob.ts";
+export * from "./expand_glob.ts";
