+++ fs/testdata/expand_wildcard.js
@@ -1,4 +1,4 @@
-import { expandGlob } from "../expand-glob.ts";
+import { expandGlob } from "../expand_glob.ts";
 
 const glob = new URL("*", import.meta.url).pathname;
 for await (const { filename } of expandGlob(glob)) {
